// src/redux/sku/types.ts
export enum ActionType {
  LOAD_CODE_TYPE = 'SKU/LOAD_CODE_TYPE',
  LOAD_CODE_TYPE_SUCCESS = 'SKU/LOAD_CODE_TYPE_SUCCESS',
  LOAD_SKU_CATEGORY = 'SKU/LOAD_SKU_CATEGORY',
  LOAD_SKU_CATEGORY_SUCCESS = 'SKU/LOAD_SKU_CATEGORY_SUCCESS',
  LOAD_SKU_BUY_NOTICE = 'SKU/LOAD_SKU_BUY_NOTICE',
  LOAD_SKU_BUY_NOTICE_SUCCESS = 'SKU/LOAD_SKU_BUY_NOTICE_SUCCESS',
  LOAD_CURRENT_SPU = 'SKU/LOAD_CURRENT_SPU',
  LOAD_CURRENT_SPU_SUCCESS = 'SKU/LOAD_CURRENT_SPU_SUCCESS',
  END_EDITING = 'SKU/END_EDITING',
}